Move utilities tests
