+++ tests/core/test_reader.py
